rename pull module to streams
This is consistent with the later taps I have
made and the naming is a little clearer this way.
It also moves the actual creation of the
Stream objects into the streams namespace,
which also helps to make things easier to find
and read, I think.
