+++ tap_xero/__init__.py
@@ -1,19 +1,11 @@
 #!/usr/bin/env python3
 import os
 import singer
+import json
 from singer import metrics, utils
 from singer.catalog import Catalog
-from .pull import (
-    IncrementingPull,
-    BankTransfersPull,
-    PaginatedPull,
-    JournalPull,
-    LinkedTransactionsPull,
-    EverythingPull,
-)
+from . import streams as streams_
 from . import credentials
-import json
-import attr
 
 CREDENTIALS_KEYS = ["consumer_key",
                     "consumer_secret",
@@ -34,59 +26,6 @@ class BadCredsException(Exception):
     pass
 
 
-@attr.attributes
-class Stream(object):
-    tap_stream_id = attr.attr()
-    pk_fields = attr.attr()
-    puller = attr.attr()
-
-STREAMS = [
-    # PAGINATED STREAMS
-    # These endpoints have all the best properties: they return the
-    # UpdatedDateUTC property and support the Modified After, order, and page
-    # parameters
-    Stream("bank_transactions", ["BankTransactionID"], PaginatedPull),
-    Stream("contacts", ["ContactID"], PaginatedPull),
-    Stream("credit_notes", ["CreditNoteID"], PaginatedPull),
-    Stream("invoices", ["InvoiceID"], PaginatedPull),
-    Stream("manual_journals", ["ManualJournalID"], PaginatedPull),
-    Stream("overpayments", ["OverpaymentID"], PaginatedPull),
-    Stream("prepayments", ["PrepaymentID"], PaginatedPull),
-    Stream("purchase_orders", ["PurchaseOrderID"], PaginatedPull),
-
-    # JOURNALS STREAM
-    # This endpoint is paginated, but in its own special snowflake way.
-    Stream("journals", ["JournalID"], JournalPull),
-
-    # NON-PAGINATED STREAMS
-    # These endpoints do not support pagination, but do support the Modified At
-    # header.
-    Stream("accounts", ["AccountID"], IncrementingPull),
-    Stream("bank_transfers", ["BankTransferID"], BankTransfersPull),
-    Stream("employees", ["EmployeeID"], IncrementingPull),
-    Stream("expense_claims", ["ExpenseClaimID"], IncrementingPull),
-    Stream("items", ["ItemID"], IncrementingPull),
-    Stream("payments", ["PaymentID"], IncrementingPull),
-    Stream("receipts", ["ReceiptID"], IncrementingPull),
-    Stream("users", ["UserID"], IncrementingPull),
-
-    # PULL EVERYTHING STREAMS
-    # These endpoints do not support the Modified After header (or paging), so
-    # we must pull all the data each time.
-    Stream("branding_themes", ["BrandingThemeID"], EverythingPull),
-    Stream("contact_groups", ["ContactGroupID"], EverythingPull),
-    Stream("currencies", ["Code"], EverythingPull),
-    Stream("organisations", ["OrganisationID"], EverythingPull),
-    Stream("repeating_invoices", ["RepeatingInvoiceID"], EverythingPull),
-    Stream("tax_rates", ["TaxType"], EverythingPull),
-    Stream("tracking_categories", ["TrackingCategoryID"], EverythingPull),
-
-    # LINKED TRANSACTIONS STREAM
-    # This endpoint is not paginated, but can do some manual filtering
-    Stream("linked_transactions", ["LinkedTransactionID"], LinkedTransactionsPull),
-]
-
-
 def get_abs_path(path):
     return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
 
@@ -105,7 +44,7 @@ def load_schema(tap_stream_id):
 
 def discover():
     result = {"streams": []}
-    for stream in STREAMS:
+    for stream in streams_.all_streams:
         schema = load_schema(stream.tap_stream_id)
         schema["selected"] = False
         result["streams"].append(
@@ -149,12 +88,11 @@ def init_credentials(config):
 def sync(config, state, catalog):
     init_credentials(config)
     currently_syncing = state.get("currently_syncing")
-    start_idx = [e.tap_stream_id for e in STREAMS].index(currently_syncing) \
-        if currently_syncing \
-        else 0
+    start_idx = streams_.all_stream_ids.index(currently_syncing) \
+        if currently_syncing else 0
     stream_ids_to_sync = [c.tap_stream_id for c in catalog.streams
                           if c.is_selected()]
-    for stream in STREAMS[start_idx:]:
+    for stream in streams_.all_streams[start_idx:]:
         if stream.tap_stream_id not in stream_ids_to_sync:
             continue
         run_stream(config, state, stream)
